energy_dashboard/app/requirements.txt
fastapi
uvicorn
requests
python-jose
passlib[bcrypt]

